requirements.txt
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.10.0
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.3.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
tzdata==2025.2
zipp==3.23.0
